Run jobs as launching user (once feature toggled on) (#97)
* Run jobs as launching user (once feature toggled on)

Currently jobs run as the group of the launching user or the group of
the proxy user (if one is configured).

This behaviour will continue until the 'use.emr.user' property is set to
true. At this point new behaviour will be enabled viz.:

If a service user is configured run as this user. If a proxy user is
configured only allow it if it is the same as the launching user. If it
is run as the prox user (not as its group).

If no proxy user or service user is configured run as the launching user
(not its group).

Utility to cycle the webserver containers added.

Utility to push dev images to dev ECR added.

Upgraded from gradle 2.

* Renamed use disallow proxy property to 'use.emr.user'

* Renamed useEmrUser property.

* Lined up test parameters.