Can check collection type dependencies.
Prior to this change the plugin could be configured to wait for a
particular product build to complete (e.g. ADG Incremental). With this
change a list of collection names can be supplied and instead of waiting
for the dependent product to be built in its entirety it can check the
UCExportToCrownStatus table for a list of collections and when those
collections have completed start the launch process.